+++ src/com/dmr/DMREmbedded.java
@@ -1,12 +1,12 @@
 package com.dmr;
 
 public class DMREmbedded {
-	private int dibit_buf[]=new int[144];
+	private byte dibit_buf[]=new byte[144];
 	private int residueValue;
 	private String line[]=new String[10];
 	private boolean resCACH,resEMB;
 	
-	public String[] decode (DMRDecode theApp,int[] buf)	{
+	public String[] decode (DMRDecode theApp,byte[] buf)	{
 		String cline;
 		DecodeCACH cachdecode=new DecodeCACH();
 		dibit_buf=buf;
